seperated with router/controller, renamed variables
