+++ backend-m2/server.js
@@ -2,6 +2,9 @@ const {
   APIError,
   calculateCarValue,
 } = require("./api/api1-takashi/carValuation");
+const {
+  calculateDiscountRate,
+} = require("./api/api4-sonny/calculateDiscountRate.js");
 const express = require("express");
 const cors = require("cors");
 require("dotenv").config();
@@ -93,36 +96,6 @@ app.post("/api/quote", (req, res) => {
 });
 
 // Sonny — API 4: Discount Rate
-
-const calculateDiscountRate = (age, yearsOfExperience) => {
-
-  // VARIABLES FOR DISCOUNT RATE CALCULATION
-  const MINIMUM_AGE_FOR_DISCOUNT = 25;
-  const MINIMUM_YEARS_OF_EXPERIENCE = 5;
-  const MAXIMUM_DISCOUNT = 20;
-  const MAXIMUM_AGE = 40;
-  const MAXIMUM_YEARS_OF_EXPERIENCE = 10;
-  const DISCOUNT_AT_FIVE_PERCENT = 5;
-  let discountRate = 0; // Start with no discount
-
-  if (age < 24 && yearsOfExperience > 5) {
-    return 0; 
-  }
-  if (age >= MINIMUM_AGE_FOR_DISCOUNT) {
-    discountRate += DISCOUNT_AT_FIVE_PERCENT;
-  }
-  if (yearsOfExperience >= MINIMUM_YEARS_OF_EXPERIENCE) {
-    discountRate += DISCOUNT_AT_FIVE_PERCENT;
-  }
-  if (age >= MAXIMUM_AGE) {
-    discountRate += DISCOUNT_AT_FIVE_PERCENT;
-  }
-  if (yearsOfExperience >= MAXIMUM_YEARS_OF_EXPERIENCE) {
-    discountRate += DISCOUNT_AT_FIVE_PERCENT;
-  }
-
-  return Math.min(discountRate, MAXIMUM_DISCOUNT);
-};
 
 // POST /test endpoint
 app.post("/test", (req, res) => {
